# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VETRA API - Platform Kesehatan Hewan Peliharaan",
    "description": "Dokumentasi Lengkap API untuk VETRA - Platform Booking Klinik, Konsultasi Dokter, Chatting, dan AI Chatbot",
    "version": "1.0.0",
    "contact": {
      "name": "Vetra API Support",
      "email": "support@vetra.com"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/chatbot/ask": {
      "post": {
        "tags": ["AI Chatbot"],
        "summary": "Bertanya ke AI Chatbot",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["message"],
                "properties": {
                  "message": { "type": "string", "example": "Kucing lemas tidak mau makan" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Respon AI dari Gemini" }
        }
      }
    },
...
    "/v2/chatbot/ask": {
      "post": {
        "tags": ["API Key Integration (v2)"],
        "summary": "Tanya AI Chatbot (API Key)",
        "security": [{"apiKeyAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["message"],
                "properties": {
                  "message": { "type": "string", "example": "Pertanyaan terintegrasi" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Respon AI sukses" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Autentikasi standar JWT. Masukkan token Anda (Tanpa menulis 'Bearer ')"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Autentikasi integrasi API Key statis. Kirimkan token API Key Anda pada header X-API-Key"
      }
    },
...
  }
}